Minor changes related with the change of name extract_unit_cell by extract_asymmetric_unit
