+++ xmipp3/protocols/__init__.py
@@ -56,7 +56,7 @@ from .protocol_eliminate_empty_images import (XmippProtEliminateEmptyParticles,
 from .protocol_extract_particles import XmippProtExtractParticles
 from .protocol_extract_particles_movies import XmippProtExtractMovieParticles
 from .protocol_extract_particles_pairs import XmippProtExtractParticlesPairs
-from .protocol_extract_unit_cell import XmippProtExtractUnit
+from .protocol_extract_asymmetric_unit import XmippProtExtractUnit
 from .protocol_helical_parameters import XmippProtHelicalParameters
 from .protocol_kerdensom import XmippProtKerdensom
 from .protocol_ml2d import XmippProtML2D
