+++ xmipp3/viewers/__init__.py
@@ -46,7 +46,7 @@ from .viewer_validate_overfitting import XmippValidateOverfittingViewer
 from .viewer_volume_strain import XmippVolumeStrainViewer
 from .viewer_reconstruct_highres import XmippReconstructHighResViewer
 from .viewer_solid_angles import SolidAnglesViewer
-from .viewer_extract_unit_cell import viewerXmippProtExtractUnit
+from .viewer_extract_asymmetric_unit import viewerXmippProtExtractUnit
 
 
 #from .viewer_combine_pdb import XmippProtCombinePdbViewer
